Initial port of Jedi Academy back to gcc/unix

+++ code/client/vmachine.h
@@ -46,7 +46,7 @@ VIRTUAL MACHINE
 ==============================================================
 */
 struct vm_s {
-	int			(*entryPoint)( int callNum, ... );
+	intptr_t			(*entryPoint)( intptr_t callNum, ... );
 };
 
 typedef struct vm_s vm_t;
@@ -54,8 +54,8 @@ typedef struct vm_s vm_t;
 extern	vm_t	cgvm;	// interface to cgame dll or vm
 extern	vm_t	uivm;	// interface to ui dll or vm
 
-extern int	VM_Call( int callnum, ... );
-extern int VM_DllSyscall( int arg, ... );
+extern intptr_t	VM_Call( intptr_t callnum, ... );
+extern intptr_t VM_DllSyscall( intptr_t arg, ... );
 extern void CL_ShutdownCGame(void);
 
 #include "../game/q_shared.h"
@@ -67,7 +67,7 @@ VM_Create
 it will attempt to load as a system dll
 ================
 */
-extern void *Sys_LoadCgame( int (**entryPoint)(int, ...), int (*systemcalls)(int, ...) );
+extern void *Sys_LoadCgame( intptr_t (**entryPoint)(intptr_t, ...), intptr_t (*systemcalls)(intptr_t, ...) );
 
 inline void *VM_Create( const char *module) 
 {
